Refactor components directory; renamed files
