+++ src/components/primitive/primitive-error/primitive-error.ts
@@ -1,10 +1,10 @@
 import {html}                    from 'lit'
 import {customElement, property} from 'lit/decorators.js'
-import {FhirElement}             from '../../../shell/FhirElement'
+import {ShoelaceStyledElement}   from '../../../shell/shoelace-styled-element'
 import {componentStyles}         from './primitive-error.styles'
 
 @customElement('fhir-error')
-export class PrimitiveError extends FhirElement {
+export class PrimitiveError extends ShoelaceStyledElement {
   static styles = componentStyles
   @property()
   declare text: string
