+++ src/components/primitive/primitive-wrapper/primitive-wrapper.ts
@@ -1,10 +1,10 @@
-import {CSSResultGroup, html} from 'lit'
-import {customElement}        from 'lit/decorators.js'
-import {FhirElement}          from '../../../shell/FhirElement'
-import {componentStyles}      from './primitive-wrapper.styles'
+import {CSSResultGroup, html}  from 'lit'
+import {customElement}         from 'lit/decorators.js'
+import {ShoelaceStyledElement} from '../../../shell/shoelace-styled-element'
+import {componentStyles}       from './primitive-wrapper.styles'
 
 @customElement('fhir-primitive-wrapper')
-export class PrimitiveWrapper extends FhirElement {
+export class PrimitiveWrapper extends ShoelaceStyledElement {
 
   static styles: CSSResultGroup = [componentStyles]
