+++ src/components/primitive/primitive.ts
@@ -11,13 +11,13 @@ import './primitive-context/primitive-context'
 import './primitive-wrapper/primitive-wrapper'
 
 import {PrimitiveType, valueOrError} from './type-converters'
-import {toCode}                      from './type-converters/ToCode'
-import {toDatetime}                  from './type-converters/ToDatetime'
-import {toDecimal}                   from './type-converters/ToDecimal'
-import {toError}                     from './type-converters/ToError'
-import {toType}                      from './type-converters/ToType'
-import {toUri}                       from './type-converters/ToUri'
-import {toUrl}                       from './type-converters/ToUrl'
+import {toCode}                      from './type-converters/toCode'
+import {toDatetime}                  from './type-converters/toDatetime'
+import {toDecimal}                   from './type-converters/toDecimal'
+import {toError}                     from './type-converters/toError'
+import {toType}                      from './type-converters/toType'
+import {toUri}                       from './type-converters/toUri'
+import {toUrl}                       from './type-converters/toUrl'
 import {asDateTime}                  from './type-presenters/asDateTime'
 import {asReadable}                  from './type-presenters/asReadable'
