+++ src/components/primitive/type-converters/index.ts
@@ -1,11 +1,11 @@
-import './ToCode'
-import './ToDatetime'
-import './ToDecimal'
-import './ToError'
-import './ToReference'
-import './ToType'
-import './ToUri'
-import './ToUrl'
+import './toCode'
+import './toDatetime'
+import './toDecimal'
+import './toError'
+import './toReference'
+import './toType'
+import './toUri'
+import './toUrl'
 
 export type toPrimitive<O, V> = (val: O) => V
 export type ValOrErr<O, V> = { type: 'ValOrErr', val: V, err: null } | { type: 'ValOrErr', val: O, err: string }
