+++ src/components/primitive/type-converters/toUri.ts
@@ -1,6 +1,6 @@
 import {URI}         from '../primitive.data'
 import {toPrimitive} from './index'
-import {toUrl}       from './ToUrl'
+import {toUrl}       from './toUrl'
 
 export const toUri: toPrimitive<string, URI> = (uri: string): URI => {
   return toUrl(uri) as URL
