+++ src/components/special/reference/reference.ts
@@ -4,7 +4,7 @@ import {customElement, state}                 from 'lit/decorators.js'
 import {choose}                               from 'lit/directives/choose.js'
 import {otherwise, when}                      from '../../.././utilities/when'
 import '../../primitive/primitive'
-import '../../../shell/layout/Wrapper'
+import '../../../shell/layout/wrapper/wrapper'
 import '../../complex/identifier/identifier'
 import {BaseElementConsumer}                  from '../../../internal/base/base-element-consumer'
 import {containedDataContext}                 from '../../../internal/contexts/context'
