+++ src/shell/index.ts
@@ -1,9 +1,8 @@
-import './layout/Wrapper'
-import './layout/empty'
-import './layout/errors'
+import './layout/wrapper/wrapper'
+import './layout/empty-set'
+import './layout/renderError'
 import './layout/not-supported'
-import './layout/structure-wrapper'
+import './layout/structure-wrapper/structure-wrapper'
 
-import './FhirElement'
-import './LoadCss'
-import './Shell'
+import './shoelace-styled-element'
+import './shell'
