--- src/shell/layout/empty.ts
@@ -1,12 +1,0 @@
-import {html}          from 'lit'
-import {customElement} from 'lit/decorators.js'
-import {FhirElement}   from '../FhirElement'
-
-@customElement('fhir-empty-set')
-export class Empty extends FhirElement {
-
-  protected render(): unknown {
-    return html`
-      <fhir-value placeholder="empty set"></fhir-value> `
-  }
-}
