+++ src/styles/hostStyles.ts
@@ -1,0 +1,17 @@
+import {css} from 'lit'
+
+export const hostStyles = css`
+  :host {
+    box-sizing: border-box;
+  }
+
+  :host *,
+  :host *::before,
+  :host *::after {
+    box-sizing: inherit;
+  }
+
+  [hidden] {
+    display: none !important;
+  }
+`
