+++ src/utilities/debug/debug.styles.ts
@@ -1,0 +1,31 @@
+import {css} from 'lit'
+
+export let componentStyles = css`
+  div {
+    padding: 0;
+    margin: 0;
+  }
+
+  ol {
+    display: inline-block;
+    background: #dddddd;
+    color: #0c2d6b;
+    border-radius: 0.3rem;
+    margin: 0.5rem;
+    padding: 0.5rem 2rem;
+  }
+
+  li {
+    font-family: monospace;
+    font-size: 0.6rem;
+    margin-top: 0.5rem;
+  }
+
+  pre {
+    margin: 0;
+  }
+
+  .key {
+    font-weight: bold;
+  }
+`
